+++ internal/version/race.go
@@ -1,0 +1,5 @@
+// +build race
+
+package version
+
+const isRace = true