+++ scripts/make/go-build.sh
@@ -54,12 +54,17 @@ esac
 # TODO(a.garipov): Additional validation?
 version="$VERSION"
 
+# Set date and time of the current build.
+buildtime="$(date -u +%FT%TZ%z)"
+
 # Set the linker flags accordingly: set the release channel and the
 # current version as well as goarm and gomips variable values, if the
 # variables are set and are not empty.
 readonly version_pkg='github.com/AdguardTeam/AdGuardHome/internal/version'
-ldflags="-s -w -X ${version_pkg}.version=${version}"
+ldflags="-s -w"
+ldflags="${ldflags} -X ${version_pkg}.version=${version}"
 ldflags="${ldflags} -X ${version_pkg}.channel=${channel}"
+ldflags="${ldflags} -X ${version_pkg}.buildtime=${buildtime}"
 if [ "${GOARM:-}" != '' ]
 then
 	ldflags="${ldflags} -X ${version_pkg}.goarm=${GOARM}"